Issue #4437: Modified CustomImportOrderCheckTest.java and moved its input files to the customimportorder subdirectory
