+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/imports/customimportorder/InputCustomImportOrderDefaultPackage.java
@@ -23,5 +23,5 @@ import com.puppycrawl.tools.*;
 import com.google.common.*;
 import org.apache.*;
 
-public class InputDefaultPackage {
+public class InputCustomImportOrderDefaultPackage {
 }
