--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/InputCustomImportOrder_NoImports.java
@@ -1,5 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
-
-
-public class InputCustomImportOrder_NoImports {
-}
