--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/InputDOMSource.java
@@ -1,7 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
-
-import javax.xml.transform.Source;
-
-import org.w3c.dom.Node;
-
-class InputDOMSource {}
