+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/customimportorder/InputCustomImportOrderPossibleIndexOutOfBoundsException.java
@@ -1,0 +1,7 @@
+package com.puppycrawl.tools.checkstyle.checks.imports.customimportorder;
+
+import javax.xml.transform.Source;
+
+import org.w3c.dom.Node;
+
+class InputCustomImportOrderPossibleIndexOutOfBoundsException {}
