+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/customimportorder/InputCustomImportOrder_NoImports.java
@@ -1,0 +1,5 @@
+package com.puppycrawl.tools.checkstyle.checks.imports.customimportorder;
+
+
+public class InputCustomImportOrder_NoImports {
+}
